chore: refactor base cog to new approach
