+++ koala/__init__.py
@@ -1,1 +1,1 @@
-from . import cogs, utils
+from . import utils
