+++ koala/cogs/__init__.py
@@ -1,0 +1,4 @@
+from .base import BaseCog
+
+from .announce import Announce
+from .twitch_alert import TwitchAlert
