+++ koala/cogs/base/__init__.py
@@ -1,0 +1,1 @@
+from .cog import BaseCog
