+++ koala/cogs/base/log.py
@@ -1,0 +1,3 @@
+from koala.log import get_logger
+
+logger = get_logger(__name__)
