+++ tests/cogs/test_BaseCog.py
@@ -20,6 +20,8 @@ from sqlalchemy import delete
 # Own modules
 import KoalaBot
 from koala.cogs import BaseCog
+from koala.cogs.base.cog import setup as setup_cog
+from koala.cogs.base.utils import new_discord_activity, list_ext_embed
 from koala.db import session_manager
 from koala.models import KoalaExtensions, GuildExtensions
 from tests.tests_utils.TestUtils import assert_activity
@@ -47,7 +49,7 @@ async def base_cog(bot):
         session.execute(delete(GuildExtensions))
         session.commit()
 
-    base_cog = BaseCog.BaseCog(bot)
+    base_cog = BaseCog(bot)
     bot.add_cog(base_cog)
     await dpytest.empty_queue()
     dpytest.configure(bot)
@@ -97,7 +99,7 @@ async def test_disable_koala_ext(base_cog):
 
 
 @pytest.mark.asyncio
-async def test_on_ready(base_cog: BaseCog.BaseCog):
+async def test_on_ready(base_cog: BaseCog):
     await base_cog.on_ready()
     assert dpytest.verify().activity().matches(discord.Activity(type=discord.ActivityType.playing,
                                              name=KoalaBot.COMMAND_PREFIX + "help" + KoalaBot.KOALA_PLUG))
@@ -118,44 +120,44 @@ async def test_invalid_change_activity():
 
 def test_playing_new_discord_activity():
     test_name = "Half Life 3"
-    assert_activity(BaseCog.new_discord_activity("playing", test_name),
+    assert_activity(new_discord_activity("playing", test_name),
                     type=discord.ActivityType.playing, name=test_name + KoalaBot.KOALA_PLUG)
 
 
 def test_watching_new_discord_activity():
     test_name = "you"
-    assert_activity(BaseCog.new_discord_activity("watching", test_name),
+    assert_activity(new_discord_activity("watching", test_name),
                     type=discord.ActivityType.watching, name=test_name + KoalaBot.KOALA_PLUG)
 
 
 def test_listening_new_discord_activity():
     test_name = "/Darude Sandstorm"
-    assert_activity(BaseCog.new_discord_activity("listening", test_name),
+    assert_activity(new_discord_activity("listening", test_name),
                     type=discord.ActivityType.listening, name=test_name + KoalaBot.KOALA_PLUG)
 
 
 def test_streaming_new_discord_activity():
     test_name = "__your room__"
-    assert_activity(BaseCog.new_discord_activity("streaming", test_name),
+    assert_activity(new_discord_activity("streaming", test_name),
                     type=discord.ActivityType.streaming, name=test_name + KoalaBot.KOALA_PLUG,
                     url=KoalaBot.STREAMING_URL)
 
 
 def test_custom_new_discord_activity():
     test_name = "1 4M K04L4"
-    assert_activity(BaseCog.new_discord_activity("custom", test_name),
+    assert_activity(new_discord_activity("custom", test_name),
                     type=discord.ActivityType.custom, name=test_name + KoalaBot.KOALA_PLUG)
 
 
 def test_invalid_new_discord_activity():
     test_name = "INCORRECT"
     with pytest.raises(SyntaxError, match="incorrect is not an activity"):
-        BaseCog.new_discord_activity("incorrect", test_name)
+        new_discord_activity("incorrect", test_name)
 
 
 @mock.patch("builtins.round", mock.MagicMock(return_value=4))
 @pytest.mark.asyncio
-async def test_ping(base_cog: BaseCog.BaseCog):
+async def test_ping(base_cog: BaseCog):
     await dpytest.message(KoalaBot.COMMAND_PREFIX + "ping")
     assert dpytest.verify().message().content("Pong! 4ms")
 
@@ -181,35 +183,35 @@ async def test_clear():
 
 
 @pytest.mark.asyncio
-async def test_invalid_clear(base_cog: BaseCog.BaseCog):
+async def test_invalid_clear(base_cog: BaseCog):
     with pytest.raises(discord.ext.commands.errors.BadArgument,
                        match="Converting to \"int\" failed for parameter \"amount\"."):
         await dpytest.message(KoalaBot.COMMAND_PREFIX + "clear a")
 
 
 @pytest.mark.asyncio
-async def test_load_cog(base_cog: BaseCog.BaseCog):
+async def test_load_cog(base_cog: BaseCog):
     with mock.patch.object(discord.ext.commands.bot.Bot, 'load_extension') as mock1:
         await dpytest.message(KoalaBot.COMMAND_PREFIX + "load_cog BaseCog")
     mock1.assert_called_with('koala.cogs.BaseCog')
 
 
 @pytest.mark.asyncio
-async def test_invalid_load_cog(base_cog: BaseCog.BaseCog):
+async def test_invalid_load_cog(base_cog: BaseCog):
     with pytest.raises(discord.ext.commands.errors.CommandInvokeError,
                        match=r".* Extension 'koala.cogs.FakeCog' could not be loaded."):
         await dpytest.message(KoalaBot.COMMAND_PREFIX + "load_cog FakeCog")
 
 
 @pytest.mark.asyncio
-async def test_unload_base_cog(base_cog: BaseCog.BaseCog):
+async def test_unload_base_cog(base_cog: BaseCog):
     with mock.patch.object(discord.ext.commands.Context, 'send') as mock1:
         await dpytest.message(KoalaBot.COMMAND_PREFIX + "unload_cog BaseCog")
     mock1.assert_called_with("Sorry, you can't unload the base cog")
 
 
 @pytest.mark.asyncio
-async def test_load_valid_cog(base_cog: BaseCog.BaseCog):
+async def test_load_valid_cog(base_cog: BaseCog):
     base_cog.COGS_DIR = "tests/tests_utils/fake_load_all_cogs"
     with mock.patch.object(discord.ext.commands.bot.Bot, 'load_extension') as mock1:
         await dpytest.message(KoalaBot.COMMAND_PREFIX + "load_cog Greetings")
@@ -217,7 +219,7 @@ async def test_load_valid_cog(base_cog: BaseCog.BaseCog):
 
 
 @pytest.mark.asyncio
-async def test_load_and_unload_valid_cog(base_cog: BaseCog.BaseCog):
+async def test_load_and_unload_valid_cog(base_cog: BaseCog):
     base_cog.COGS_DIR = "tests/tests_utils/fake_load_all_cogs"
     with mock.patch.object(discord.ext.commands.bot.Bot, 'load_extension') as mock1:
         await dpytest.message(KoalaBot.COMMAND_PREFIX + "load_cog Greetings")
@@ -229,7 +231,7 @@ async def test_load_and_unload_valid_cog(base_cog: BaseCog.BaseCog):
 
 
 @pytest.mark.asyncio
-async def test_invalid_unload_cog(base_cog: BaseCog.BaseCog):
+async def test_invalid_unload_cog(base_cog: BaseCog):
     with pytest.raises(discord.ext.commands.errors.CommandInvokeError,
                        match="Command raised an exception: ExtensionNotLoaded:"
                              " Extension 'koala.cogs.FakeCog' has not been loaded."):
@@ -237,7 +239,7 @@ async def test_invalid_unload_cog(base_cog: BaseCog.BaseCog):
 
 
 @pytest.mark.asyncio
-async def test_version(base_cog: BaseCog.BaseCog):
+async def test_version(base_cog: BaseCog):
     await dpytest.message(KoalaBot.COMMAND_PREFIX + "version")
     assert dpytest.verify().message().content("version: "+KoalaBot.__version__)
 
@@ -246,5 +248,5 @@ async def test_version(base_cog: BaseCog.BaseCog):
 @pytest.mark.asyncio
 async def test_setup():
     with mock.patch.object(discord.ext.commands.bot.Bot, 'add_cog') as mock1:
-        BaseCog.setup(KoalaBot.client)
+        setup_cog(KoalaBot.client)
     mock1.assert_called()
